共通した関数をuse以下へ
